feat: Improve CDN origin control (#140)
* feat: Enable explicit origin type definiton

The module is meant to be used with custom origins, but it may happen
that your origin of choice is an S3 bucket that's not managed by
yourself. `cloudposse/cloudfront-s3-cdn/aws` may seem to be a viable
alternative (it supports pre-existing buckets), but there's an implicit
assumption there - both the origin bucket and the CloudFront
distribution that points to it should be deployed to the same AWS
account. In order to secure CDN-to-S3 communication one may want to
specify either OAI or OAC. Prior to this change, it was not possible to
use OAI when S3 bucket domain name was used.

* fix: Set OAC only when S3 type origin was used

When user specifies an S3 domain name as the origin and attaches OAC to
it, Terraform fails with the following error:

Illegal configuration: The origin type and OAC origin type differ.

* feat: Make sure origin shield can be specified for non-default origins

At the moment origin shield can be configured only for the default
origin. This commit enables the feature for all the remaining ones.

* fix: Replace unsupported map() references

`map()` function is no longer available

* fix: Zero all TTL params when cache policy ID is configured

When cache policy ID is specified, all TTL arguments should be zeroed,
as the policy-scoped TTLs take precedence. If non-zero `min_ttl`,
`default_ttl` or `max_ttl` is set to non-0 value and at the same time
cache policy is specified it leads to phantom changes in the plan:

default_ttl                = 0 -> 60

* Revert "fix: Zero all TTL params when cache policy ID is configured"

This reverts commit 30ff7b9dde2d28ad91850406247ac11c3bf639e3.

This should be moved to a separate PR to narrow down the scope just to
origin-related changes

* docs: Reflect new/updated variables in the README file

* fix: Fallback to default OAI when S3 origin was specified without `s3_origin_config` variable

* feat: Add S3-specific examples

* chore: Add an explicit comment to OAI resource

* chore: Update OAI name

* chore: Update s3_origin_config variable description

* feat: Init s3_origin_config block only when OAC wasn't provided

* fix: Improve S3 examples

* chore: Resource naming convention

---------

Co-authored-by: Dan Miller <miller0daniel@gmail.com>